commands: move duplicate code to duplicate.rs
